Turn VariableRole into an enumerated type and move to graph.py
